Move define.TempDirForURL to a separate subpackage
It's surprising to have it in a package that otherwise
contains almost exclusively type definitions, and we are
breaking the API either way.

Should not change behavior.
